fix: handle animations where totalFrames < fps

+++ camcorder/src/jvmMain/kotlin/xyz/schwaab/camcorder/AnimationSpecs.kt
@@ -18,7 +18,7 @@ data class AnimationSpecs(
     ): this(
         width,
         height,
-        (totalFrames / fps).seconds,
+        (totalFrames.toDouble() / fps).seconds,
         fps
     )
